backend/requirements.txt
django==5.2
djangorestframework==3.16.0
djoser==2.3.1
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
pillow==11.2.1
python-dotenv==1.0.1 